# app/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "app, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json;odata.metadata=minimal;odata.streaming=true",
          "application/json;odata.metadata=minimal;odata.streaming=false",
          "application/json;odata.metadata=minimal",
          "application/json;odata.metadata=full;odata.streaming=true",
          "application/json;odata.metadata=full;odata.streaming=false",
          "application/json;odata.metadata=full",
          "application/json;odata.metadata=none;odata.streaming=true",
          "application/json;odata.metadata=none;odata.streaming=false",
          "application/json;odata.metadata=none",
          "application/json;odata.streaming=true",
          "application/json;odata.streaming=false",
          "application/json",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=full;IEEE754Compatible=false",
          "application/json;odata.metadata=full;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=true",
          "application/json;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.streaming=false;IEEE754Compatible=true",
          "application/json;IEEE754Compatible=false",
          "application/json;IEEE754Compatible=true",
          "application/xml",
          "text/plain",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json;odata.metadata=minimal;odata.streaming=true",
          "application/json;odata.metadata=minimal;odata.streaming=false",
          "application/json;odata.metadata=minimal",
          "application/json;odata.metadata=full;odata.streaming=true",
          "application/json;odata.metadata=full;odata.streaming=false",
          "application/json;odata.metadata=full",
          "application/json;odata.metadata=none;odata.streaming=true",
          "application/json;odata.metadata=none;odata.streaming=false",
          "application/json;odata.metadata=none",
          "application/json;odata.streaming=true",
          "application/json;odata.streaming=false",
          "application/json",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=full;IEEE754Compatible=false",
          "application/json;odata.metadata=full;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=true",
          "application/json;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.streaming=false;IEEE754Compatible=true",
          "application/json;IEEE754Compatible=false",
          "application/json;IEEE754Compatible=true",
          "application/xml",
          "text/plain",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/auth/forgot_password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json;odata.metadata=minimal;odata.streaming=true",
          "application/json;odata.metadata=minimal;odata.streaming=false",
          "application/json;odata.metadata=minimal",
          "application/json;odata.metadata=full;odata.streaming=true",
          "application/json;odata.metadata=full;odata.streaming=false",
          "application/json;odata.metadata=full",
          "application/json;odata.metadata=none;odata.streaming=true",
          "application/json;odata.metadata=none;odata.streaming=false",
          "application/json;odata.metadata=none",
          "application/json;odata.streaming=true",
          "application/json;odata.streaming=false",
          "application/json",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=full;IEEE754Compatible=false",
          "application/json;odata.metadata=full;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=true",
          "application/json;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.streaming=false;IEEE754Compatible=true",
          "application/json;IEEE754Compatible=false",
          "application/json;IEEE754Compatible=true",
          "application/xml",
          "text/plain",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ForgotPasswordForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/auth/reset_password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json;odata.metadata=minimal;odata.streaming=true",
          "application/json;odata.metadata=minimal;odata.streaming=false",
          "application/json;odata.metadata=minimal",
          "application/json;odata.metadata=full;odata.streaming=true",
          "application/json;odata.metadata=full;odata.streaming=false",
          "application/json;odata.metadata=full",
          "application/json;odata.metadata=none;odata.streaming=true",
          "application/json;odata.metadata=none;odata.streaming=false",
          "application/json;odata.metadata=none",
          "application/json;odata.streaming=true",
          "application/json;odata.streaming=false",
          "application/json",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=full;IEEE754Compatible=false",
          "application/json;odata.metadata=full;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=true",
          "application/json;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.streaming=false;IEEE754Compatible=true",
          "application/json;IEEE754Compatible=false",
          "application/json;IEEE754Compatible=true",
          "application/xml",
          "text/plain",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ResetPasswordForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ForgotPasswordForm": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "LoginForm": {
      "type": "object",
      "properties": {
        "emailOrUsername": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "remember": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "RegisterForm": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "ResetPasswordForm": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}